Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float x;
-    scanf("%.4f", &x);
-    printf("You entered: %f", x);
+    double x;
+    scanf("%lf", &x);
+    printf("You entered: %lf", x);
     return 0;
 }
